refactor(signal): isolate outbound encoding and consolidate attachment proof
