+++ extensions/signal/src/client-container.real-server.test.ts
@@ -129,6 +129,7 @@ describe("signal REST real-server deadline", () => {
   });
 
   it.each([
+    { stagedFilename: "test-image.jpg", expectedFilename: "test-image.jpg" },
     {
       stagedFilename: "report---a1b2c3d4-5678-90ab-cdef-1234567890ab.jpg",
       expectedFilename: "report.jpg",
