+++ extensions/signal/src/client-container.test.ts
@@ -894,117 +894,6 @@ describe("containerSendMessage", () => {
     expect(body.message).toBe("**Bold** C:\\Temp\\file and /foo\\bar/");
   });
 
-  it("includes attachments as base64 data URIs", async () => {
-    const fs = await import("node:fs/promises");
-    const os = await import("node:os");
-    const path = await import("node:path");
-
-    // Create a temp file with known content
-    const tmpDir = await fs.mkdtemp(path.join(os.tmpdir(), "signal-test-"));
-    const tmpFile = path.join(tmpDir, "test-image.jpg");
-    const content = Buffer.from([0xff, 0xd8, 0xff, 0xe0]); // JPEG magic bytes
-    await fs.writeFile(tmpFile, content);
-
-    mockFetch.mockResolvedValue({
-      ok: true,
-      status: 200,
-      ...bodyStream(JSON.stringify({})),
-    });
-
-    await containerSendMessage({
-      baseUrl: "http://localhost:8080",
-      account: "+14259798283",
-      recipients: ["+15550001111"],
-      message: "Photo",
-      attachments: [tmpFile],
-    });
-
-    const body = parseFetchBody();
-    expect(body.attachments).toBeUndefined();
-    if (!Array.isArray(body.base64_attachments)) {
-      throw new Error("expected base64 attachments array");
-    }
-    expect(body.base64_attachments).toHaveLength(1);
-    expect(body.base64_attachments[0]).toMatch(
-      /^data:image\/jpeg;filename=test-image\.jpg;base64,/,
-    );
-
-    // Cleanup
-    await fs.rm(tmpDir, { recursive: true });
-  });
-
-  it.each([
-    {
-      stagedFilename: "report---a1b2c3d4-5678-90ab-cdef-1234567890ab.jpg",
-      expectedFilename: "report.jpg",
-    },
-    {
-      stagedFilename: "quarter;final---a1b2c3d4-5678-90ab-cdef-1234567890ab.jpg",
-      expectedFilename: "quarter_final.jpg",
-    },
-    {
-      stagedFilename: "first;middle;last---a1b2c3d4-5678-90ab-cdef-1234567890ab.jpg",
-      expectedFilename: "first_middle_last.jpg",
-    },
-    {
-      stagedFilename: "quarter,final---a1b2c3d4-5678-90ab-cdef-1234567890ab.jpg",
-      expectedFilename: "quarter_final.jpg",
-    },
-    {
-      stagedFilename: "first,middle,last---a1b2c3d4-5678-90ab-cdef-1234567890ab.jpg",
-      expectedFilename: "first_middle_last.jpg",
-    },
-    {
-      stagedFilename: "hash#name---a1b2c3d4-5678-90ab-cdef-1234567890ab.jpg",
-      expectedFilename: "hash_name.jpg",
-    },
-    {
-      stagedFilename: "mixed;comma,hash#name---a1b2c3d4-5678-90ab-cdef-1234567890ab.jpg",
-      expectedFilename: "mixed_comma_hash_name.jpg",
-    },
-    { stagedFilename: "quarter;final.jpg", expectedFilename: "quarter_final.jpg" },
-    { stagedFilename: "quarter,final.jpg", expectedFilename: "quarter_final.jpg" },
-    { stagedFilename: "hash#name.jpg", expectedFilename: "hash_name.jpg" },
-    {
-      stagedFilename: "quarter final---a1b2c3d4-5678-90ab-cdef-1234567890ab.jpg",
-      expectedFilename: "quarter final.jpg",
-    },
-  ])(
-    "restores the provider-safe original attachment filename $expectedFilename",
-    async ({ stagedFilename, expectedFilename }) => {
-      const fs = await import("node:fs/promises");
-      const os = await import("node:os");
-      const path = await import("node:path");
-
-      const tmpDir = await fs.mkdtemp(path.join(os.tmpdir(), "signal-test-"));
-      try {
-        const stagedFile = path.join(tmpDir, stagedFilename);
-        const content = Buffer.from([0xff, 0xd8, 0xff, 0xe0]);
-        await fs.writeFile(stagedFile, content);
-
-        mockFetch.mockResolvedValue({
-          ok: true,
-          status: 200,
-          ...bodyStream(JSON.stringify({})),
-        });
-
-        await containerSendMessage({
-          baseUrl: "http://localhost:8080",
-          account: "+14259798283",
-          recipients: ["+15550001111"],
-          message: "Photo",
-          attachments: [stagedFile],
-        });
-
-        expect(parseFetchBody().base64_attachments).toEqual([
-          `data:image/jpeg;filename=${expectedFilename};base64,${content.toString("base64")}`,
-        ]);
-      } finally {
-        await fs.rm(tmpDir, { recursive: true, force: true });
-      }
-    },
-  );
-
   it("rejects outbound attachments that exceed the size cap", async () => {
     const fs = await import("node:fs/promises");
     const os = await import("node:os");
